Update project references, SQL scripts, and data types
- Increment package version in `Package.wxs` to 1.0.0.16.
- Modify SQL scripts to enhance delete logic and add new conditions.
- Update SQL functions to use `decimal` for weight calculations.
- Change parameter types in stored procedures from `float` to `decimal`.
- Update DTO classes in `ItemDTO.cs` and `MaterialDTO.cs` for better precision.
- Enhance `ProgressValue` class for improved progress tracking.
- Revise `WriteService` and `PrefSuiteService` for better progress reporting.
- Update UI components in WinForms for improved user experience.
- Improve error logging in `ExcelService` when writing to Excel files.

+++ src/a2p.DataAssets/SQL/Functions.sql
@@ -133,18 +133,18 @@ CREATE OR ALTER FUNCTION [dbo].[Uniwave_a2p_GetTechDesignCommodityCode]
 RETURNS int
 AS
 BEGIN
-    DECLARE @CommodityCode Nvarchar(25), 
-             @Code INT 
-     
-    SELECT TOP 1 @CommodityCode =  [Default commodity code]
-    FROM Nordan_a2p_IntrastatData WHERE 
-    CASE 
-        WHEN CHARINDEX('.', Material) > 0 
-        THEN LEFT(Material, CHARINDEX('.', Material) - 1)
-        ELSE Material
-    END  = @SourceReference
-    
-    SELECT TOP 1 @Code= Id FROM Intrastat.CommodityCodes Where Code = @CommodityCode
+	DECLARE @CommodityCode Nvarchar(25), 
+			 @Code INT 
+	 
+	SELECT TOP 1 @CommodityCode =  [Default commodity code]
+	FROM Nordan_a2p_IntrastatData WHERE 
+	CASE 
+		WHEN CHARINDEX('.', Material) > 0 
+		THEN LEFT(Material, CHARINDEX('.', Material) - 1)
+		ELSE Material
+	END  = @SourceReference
+	
+	SELECT TOP 1 @Code= Id FROM Intrastat.CommodityCodes Where Code = @CommodityCode
 
 
 RETURN @Code
@@ -157,19 +157,19 @@ CREATE OR ALTER FUNCTION [dbo].[Uniwave_a2p_GetTechDesignWeight]
 (
  @SourceReference Nvarchar(50)
 )
-RETURNS Float
+RETURNS decimal
 AS
 BEGIN
-    DECLARE @Weight Float 
-    SELECT TOP 1 @Weight = Weight
-    FROM Nordan_a2p_IntrastatData NINT
-    WHERE 
+	DECLARE @Weight decimal (38,6) 
+	SELECT TOP 1 @Weight = Weight
+	FROM Nordan_a2p_IntrastatData NINT
+	WHERE 
 
-    CASE 
-        WHEN CHARINDEX('.', NINT.Material) > 0 
-        THEN LEFT(NINT.Material, CHARINDEX('.', NINT.Material) - 1)
-        ELSE NINT.Material
-    END =  @SourceReference
+	CASE 
+		WHEN CHARINDEX('.', NINT.Material) > 0 
+		THEN LEFT(NINT.Material, CHARINDEX('.', NINT.Material) - 1)
+		ELSE NINT.Material
+	END =  @SourceReference
 
 RETURN @Weight
